# ui/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "joinedcontext Portal API",
    "description": "Administrative and platform management REST API for joinedcontext Portal",
    "license": {
      "name": "EUPL-1.2",
      "identifier": "EUPL-1.2"
    },
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/projects/{project}/agent-runs/{id}/preview": {
      "get": {
        "tags": [
          "agents"
        ],
        "summary": "The preview of a kit run: the bundle, the specification and the stylesheet in one document,\nbecause the frame it is shown in has no origin to fetch anything else with (AP-60, UI-41).",
        "operationId": "preview",
        "parameters": [
          {
            "name": "project",
            "in": "path",
            "description": "Project name",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "id",
            "in": "path",
            "description": "Run identifier",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The kit rendering the run's specification, one document",
            "content": {
              "text/html": {}
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "404": {
            "description": "No such run, or no pass has written spec.json yet",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "This Portal was built without the kit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/tools/generate": {
      "post": {
        "tags": [
          "tools"
        ],
        "operationId": "generate",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compiled artifacts, or the messages of a source that does not compile",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifacts"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "413": {
            "description": "Source larger than the payload limit",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No model tools service configured, or it did not answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/import-sdm": {
      "post": {
        "tags": [
          "tools"
        ],
        "operationId": "import_sdm",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ImportSdmRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Compiled artifacts of the catalogue model",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Artifacts"
                }
              }
            }
          },
          "400": {
            "description": "Not a Smart Data Models catalogue identifier",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No model tools service configured, or it did not answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/tools/sdm-catalog": {
      "get": {
        "tags": [
          "tools"
        ],
        "operationId": "sdm_catalog",
        "parameters": [
          {
            "name": "refresh",
            "in": "query",
            "description": "Refresh the cached index now instead of waiting for the daily run",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The catalogue index, from the cache when a refresh did not reach the catalogue",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Catalogue"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          },
          "503": {
            "description": "No model tools service configured, or it did not answer",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Artifacts": {
        "type": "object",
        "description": "What one Model Tools run rendered. Every artifact is optional: a version that does not\nrender one omits it, and a source that does not compile yet answers with `errors` filled in\nand the artifacts absent. A half-written model is the normal state of an editor.",
        "properties": {
          "context": {},
          "docs": {
            "type": [
              "string",
              "null"
            ],
            "description": "The documentation page, one Markdown file for the whole model: the same page\n`jcctl model generate` commits beside the source, so the editor previews what a\nreviewer will approve rather than a second rendering of it (DM-02, DM-32, DM-43)."
          },
          "errors": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Compilation messages. Non-empty with no artifacts means the source does not compile."
          },
          "example": {},
          "generatorVersion": {
            "type": [
              "string",
              "null"
            ],
            "description": "The generator version that produced these artifacts, so a preview and a committed\nartifact set can be compared (DM-19)."
          },
          "jsonSchema": {},
          "linkml": {
            "type": [
              "string",
              "null"
            ],
            "description": "The LinkML source itself, which an import produces and the editor then edits; its\nannotations carry `spec.source.repository`, `path` and `commit` (DM-07, DM-08)."
          },
          "owl": {
            "type": [
              "string",
              "null"
            ]
          },
          "shacl": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "Catalogue": {
        "type": "object",
        "description": "The catalogue index Model Tools caches and refreshes daily (DM-12).",
        "properties": {
          "refreshedAt": {
            "type": [
              "string",
              "null"
            ],
            "description": "When the cache was last filled from the catalogue."
          },
          "stale": {
            "type": "boolean",
            "description": "The last refresh did not reach the catalogue, so this index is the older cached one.\nAn index a browser cannot refresh is still an index it can work from (DM-12)."
          },
          "subjects": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogueSubject"
            }
          }
        }
      },
      "CatalogueModel": {
        "type": "object",
        "description": "One model of the Smart Data Models catalogue index, as the wizard lists and searches it.",
        "required": [
          "id",
          "name"
        ],
        "properties": {
          "attributes": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Attribute names, so the wizard can search by attribute without fetching the model."
          },
          "description": {
            "type": [
              "string",
              "null"
            ]
          },
          "id": {
            "type": "string",
            "description": "The catalogue identifier, `dataModel.Environment/AirQualityObserved`."
          },
          "name": {
            "type": "string",
            "description": "The model name, `AirQualityObserved`."
          }
        }
      },
      "CatalogueSubject": {
        "type": "object",
        "description": "One subject of the catalogue: `dataModel.Environment` and the models under it.",
        "required": [
          "name"
        ],
        "properties": {
          "models": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CatalogueModel"
            }
          },
          "name": {
            "type": "string"
          },
          "title": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
      "GenerateRequest": {
        "type": "object",
        "description": "A LinkML source to compile.",
        "required": [
          "source"
        ],
        "properties": {
          "source": {
            "type": "string",
            "description": "LinkML YAML, exactly as the editor holds it."
          }
        }
      },
...
      "ImportSdmRequest": {
        "type": "object",
        "description": "A model to import from the Smart Data Models catalogue.",
        "required": [
          "model"
        ],
        "properties": {
          "model": {
            "type": "string",
            "description": "Catalogue identifier, `dataModel.Environment/AirQualityObserved`. Never a URL: the\nallowlist that limits fetching to the Smart Data Models organisation lives in Model\nTools (DM-10), and the Portal refuses anything a caller could steer."
          }
        }
      },
...
      "ProblemDetails": {
        "type": "object",
        "description": "RFC 7807 Problem Details representation.",
        "required": [
          "type",
          "title",
          "status"
        ],
        "properties": {
          "detail": {
            "type": [
              "string",
              "null"
            ]
          },
          "errors": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            },
            "description": "Every violation of one request, when there is more than one thing to say. A form marks\nall its bad fields from this in one pass instead of one round trip per mistake (CC-24)."
          },
          "instance": {
            "type": [
              "string",
              "null"
            ]
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "title": {
            "type": "string"
          },
          "type": {
            "type": "string"
          }
        }
      },
...
    }
  },
...
}